requirements.txt
anthropic>=0.40.0
pymupdf>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.24.0
